requirements.txt
python-dotenv
disnake
dsplayer==2.2.0